requirements.txt
flask==3.1.3
gunicorn==25.3.0
langchain==1.2.15
langchain-community==0.4.1
langchain-openai==1.2.1
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.2
openai==2.32.0
pinecone-client==6.0.0
pypdf==6.10.2
python-dotenv==1.2.2
tiktoken==0.12.0
numpy==1.26.4
